Refactor - rename file, move logic to single file (a bit more cohesive).
